requirements.txt
# 核心依赖
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
pyannote.audio==3.1.1

# 音频处理
librosa>=0.10.0
pydub>=0.25.1
soundfile>=0.12.1
noisereduce>=2.0.0

# Web框架（可选，用于API）
fastapi>=0.104.0
uvicorn>=0.24.0

# 异步处理
celery>=5.3.0
redis>=5.0.0

# 数据处理
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0

# 工具类
tqdm>=4.65.0
loguru>=0.7.0

# 开发工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# NLP处理
jieba>=0.42.1
dateparser>=1.2.0
keybert>=0.8.0
transformers>=4.40.0

# 大模型
openai>=1.30.0
qianfan>=0.3.0  # 百度千帆
dashscope>=1.14.0  # 阿里通义千问

# 可视化
matplotlib>=3.7.0
plotly>=5.22.0
jinja2>=3.1.0
pandas>=2.0.0

# 其他
pydantic>=2.0.0
python-dateutil>=2.8.0
